+++ server/src/entity/Professor.ts
@@ -1,4 +1,4 @@
-import { Field, ID, Int, ObjectType, Root } from 'type-graphql';
+import { Field, Float, ID, Int, ObjectType, Root } from 'type-graphql';
 import { BaseEntity, Column, Entity, OneToMany, PrimaryGeneratedColumn } from 'typeorm';
 import { Comment } from './Comment';
 
@@ -25,7 +25,7 @@ export class Professor extends BaseEntity {
     @Column('int', { array: true })
     difficulty: number[] = [];
 
-    @Field(() => Int, { nullable: true })
+    @Field(() => Float, { nullable: true })
     averageDifficulty(@Root() parent: Professor): number | null {
         if (parent.difficulty.length) {
             return (
@@ -41,7 +41,7 @@ export class Professor extends BaseEntity {
     @Column('int', { array: true })
     quality: number[] = [];
 
-    @Field(() => Int, { nullable: true })
+    @Field(() => Float, { nullable: true })
     averageQuality(@Root() parent: Professor): number | null {
         if (parent.quality.length) {
             return (
